domain_tools/requirements.txt
aiohttp==3.8.4
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
frozenlist==1.2.0
idna==3.4
idna-ssl==1.1.0
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
PyYAML==6.0
requests==2.27.1
typing_extensions==4.1.1
urllib3==1.26.15
uvloop==0.14.0
yarl==1.7.2
zipp==3.6.0
